requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alembic=1.11.1=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2023.05.30=haa95532_0
certifi=2023.7.22=py38haa95532_0
cffi=1.15.1=py38h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cmaes=0.10.0=pypi_0
colorama=0.4.6=pypi_0
colorlog=6.7.0=pypi_0
contourpy=1.1.0=pypi_0
cpuonly=2.0=0
cryptography=41.0.2=py38hac1b9e3_0
cycler=0.11.0=pypi_0
dhg=0.9.3=pypi_0
fonttools=4.41.1=pypi_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
greenlet=2.0.2=pypi_0
idna=3.4=py38haa95532_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.0=pypi_0
intel-openmp=2023.1.0=h59b6b97_46319
joblib=1.3.1=pypi_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=pypi_0
lerc=3.0=hd77b12b_0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
mako=1.2.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
metric-learn=0.6.2=pypi_0
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.6=py38hf11a4ad_1
mkl_random=1.2.2=py38hf11a4ad_1
networkx=3.1=pypi_0
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
openssl=3.0.9=h2bbff1b_0
optuna=3.2.0=pypi_0
packaging=23.1=pypi_0
pandas=1.5.0=pypi_0
pillow=9.4.0=py38hd77b12b_0
pip=23.2.1=py38haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py38haa95532_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38haa95532_0
python=3.8.17=h1aa4202_0
python-dateutil=2.8.2=pypi_0
pytorch=1.12.1=py3.8_cpu_0
pytorch-mutex=1.0=cpu
pytz=2023.3=pypi_0
pyyaml=6.0.1=pypi_0
requests=2.29.0=py38haa95532_0
scikit-learn=1.3.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.0.0=py38haa95532_0
six=1.16.0=pypi_0
sqlalchemy=2.0.19=pypi_0
sqlite=3.41.2=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h2bbff1b_0
torchaudio=0.12.1=py38_cpu
torchvision=0.13.1=py38_cpu
tqdm=4.65.0=pypi_0
typing_extensions=4.7.1=py38haa95532_0
urllib3=1.26.16=py38haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.4.2=h8cc25b3_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
